GrammarCorrecter/requirements.txt
streamlit==1.31.0
transformers==4.41.2
torch==2.2.1
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.26.3
sentencepiece==0.1.99
protobuf==4.25.1